Avoid double download

+++ runtime-install.sh
@@ -39,7 +39,10 @@ pip3 install -U "setuptools<50" pip "cython<3" "wheel==0.42.0" && pip3 install g
 
 # Tensorflow and dependencies
 cd $DIR/dist/
-gdown --fuzzy https://drive.google.com/file/d/1rfgF2U2oZJvQSMbGNZl8f5jbWP4fY6UW/view?usp=sharing
+if [ ! -f "$DIR/dist/tensorflow-2.4.1-cp38-cp38-linux_aarch64.wh" ]; then
+    echo "Downloading TensorFlow wheel exists."
+    gdown --fuzzy https://drive.google.com/file/d/1rfgF2U2oZJvQSMbGNZl8f5jbWP4fY6UW/view?usp=sharing
+fi
 pip3 install pyudev \
     "flask<3" \
     flask_cors \
@@ -60,14 +63,15 @@ echo "DNSStubListener=no" |  tee -a /etc/systemd/resolved.conf >/dev/null
 systemctl restart systemd-resolved
 
 # Install packages
-curl -O https://larsll-build-artifact-share.s3.eu-north-1.amazonaws.com/deepracer-pi/aws-deepracer-core_2.0.383.1%2Bcommunity_arm64.deb
-curl -O https://larsll-build-artifact-share.s3.eu-north-1.amazonaws.com/deepracer-pi/aws-deepracer-device-console_2.0.196.0_arm64.deb
-curl -O https://larsll-build-artifact-share.s3.eu-north-1.amazonaws.com/deepracer-pi/aws-deepracer-sample-models_2.0.9.0_all.deb
-curl -O https://larsll-build-artifact-share.s3.eu-north-1.amazonaws.com/deepracer-pi/aws-deepracer-util_2.0.61.0_arm64.deb
+cd $DIR/dist/
+[ ! -f "$DIR/dist/aws-deepracer-core_2.0.383.1%2Bcommunity_arm64.deb" ] && curl -O https://larsll-build-artifact-share.s3.eu-north-1.amazonaws.com/deepracer-pi/aws-deepracer-core_2.0.383.1%2Bcommunity_arm64.deb
+[ ! -f "$DIR/dist/aws-deepracer-device-console_2.0.196.0_arm64.deb" ] && curl -O https://larsll-build-artifact-share.s3.eu-north-1.amazonaws.com/deepracer-pi/aws-deepracer-device-console_2.0.196.0_arm64.deb
+[ ! -f "$DIR/dist/aws-deepracer-sample-models_2.0.9.0_all.deb" ] && curl -O https://larsll-build-artifact-share.s3.eu-north-1.amazonaws.com/deepracer-pi/aws-deepracer-sample-models_2.0.9.0_all.deb
+[ ! -f "$DIR/dist/aws-deepracer-util_2.0.61.0_arm64.deb" ] && curl -O https://larsll-build-artifact-share.s3.eu-north-1.amazonaws.com/deepracer-pi/aws-deepracer-util_2.0.61.0_arm64.deb
 apt install -y ./*.deb
 
 # Get OpenVINO
-curl -O https://larsll-build-artifact-share.s3.eu-north-1.amazonaws.com/deepracer-pi/openvino_2021.3_arm64.tgz
+[ ! -f "$DIR/dist/openvino_2021.3_arm64.tgz" ] && curl -O https://larsll-build-artifact-share.s3.eu-north-1.amazonaws.com/deepracer-pi/openvino_2021.3_arm64.tgz
 cd /
 tar xvzf $DIR/distn/openvino_2021.3_arm64.tgz
 ln -sf /opt/intel/openvino_2021.3 /opt/intel/openvino_2021
